+++ packages/dara-components/changelog.md
@@ -4,7 +4,7 @@ title: Changelog
 
 ## NEXT
 
--  Fixed an issue where `CodeEditor` would bubble up Enter keypresses when confirming composition using IME input
+-  Fixed an issue with `CodeEditor`, `Input`, `Select`, `Textarea` components where they would bubble up Enter keypresses when confirming composition using IME input
 
 ## 1.16.17